move handleLogout from dashboard to profile nav
