+++ src/components/navbars/DesktopNav.jsx
@@ -1,9 +1,12 @@
+import ProfileNav from './ProfileNav';
 import loginIcon from '../../assets/icons/login.png';
 import searchIcon from '../../assets/icons/search.png';
+import { useAuth } from '../../AuthProvider';
 
 const DesktopNav = ({ onSearch, onLogin, onJoin, stateSearch, stateLogin, stateJoin }) => {
+    const { isLoggedIn } = useAuth()
     return (
-        <nav className="w-2/3 hidden relative py-6 sm:flex justify-between">
+        <nav className="w-2/3 hidden relative py-6 sm:flex justify-between items-center">
             {
                 stateSearch
                     ? <form className="border-2 border-red rounded-full h-10 w-2/4 flex justify-between items-center px-4 ml-6 gap-1">
@@ -13,20 +16,24 @@ const DesktopNav = ({ onSearch, onLogin, onJoin, stateSearch, stateLogin, stateJ
                     : <div></div>
             }
             <div id="mainNav" className="flex gap-4 items-center">
-                {
-                    stateLogin
-                        ? <button onClick={onLogin} className="font-spartan font-semibold text-base sm:text-lg flex items-center">
-                            <img src={loginIcon} className="h-4 sm:h-5" alt="join icon" />
-                            Log In
-                        </button>
-                        : <div></div>
-                }
-                {
-                    stateJoin
-                        ? <button onClick={onJoin} className='bg-red hover:bg-pureWhite hover:text-red w-16 h-8 hover:border-2 hover:border-red text-white font-spartan font-semibold text-base sm:text-lg rounded-lg transition duration-300 easy-in'>
-                            Join
-                        </button>
-                        : <div></div>
+                {isLoggedIn
+                    ? <ProfileNav />
+                    : <>
+                        {stateLogin
+                            ? <button onClick={onLogin} className="font-spartan font-semibold text-base sm:text-lg flex items-center">
+                                <img src={loginIcon} className="h-4 sm:h-5" alt="join icon" />
+                                Log In
+                            </button>
+                            : <div></div>
+                        }
+                        {
+                            stateJoin
+                                ? <button onClick={onJoin} className='bg-red hover:bg-pureWhite hover:text-red w-16 h-8 hover:border-2 hover:border-red text-white font-spartan font-semibold text-base sm:text-lg rounded-lg transition duration-300 easy-in'>
+                                    Join
+                                </button>
+                                : <div></div>
+                        }
+                    </>
                 }
             </div>
         </nav>)
